+++ src/main/java/ru/yandex/practicum/filmorate/storage/genre/GenreStorage.java
@@ -5,6 +5,6 @@ import ru.yandex.practicum.filmorate.model.Genre;
 import java.util.List;
 
 public interface GenreStorage {
-    Genre get(int id);
+    Genre get(Long id);
     List<Genre> getAll();
 }
